Fixes to jewelry patch
Previously, this was excluding from dq_mki, dq_mkii, and billtypeminer_i. I _think_ the intent was to exclude from all and give them a small chance, based on tier, of getting those valuable gemstone ores (I inferred this from the miners XML). If so, this is a good patch to do so. If I misunderstood the intent, it is not. Let me know and I can correct this patch and resubmit.

Also of note, `starts-with(defName, "PRF_DeepQuarry_mk") or defName="PRF_BillTypeMiner_I"` lets you save 3 `PatchOperationAdd` nodes. Not going to have much performance impact but applied to a large enough patching project, the pattern could add up.