Patch for Oriental Dragon mod.
Based on the WhatTheHack patch.  "Syulipia_Gate_Dragon" was being taken as a valid ore to be mined, and so it would either obliterate the miner as someone mentioned or it would straight-up crash the game if you used the highest tier miner to create one using a bill deliberately.