Update Recipes_Mining.xml
So, the base recipe here has a work defined of "500" but then every single instance where it's used in the base mod it gets overwritten with a work value of 655.  The weird consequence of this is that any patches which add additional stones using the base that don't manually define a work value of 655 will be easier to mine than vanilla stone by 25-30%.

I recommend just removing the manual overrides and changing the work to 655 in the base definition.