Update Scenarios_PRF.xml
Remove the two Wood Fired generators from the scenario and replace them with an equivalent amount of resources (200 steel and 4 components). Because these generators are not minifiable in an otherwise unmodded game, they spawn already set up exactly wherever the colonists arrive, which means they might just pop up in the middle of a swamp or somewhere else useless.