Update AllKeyed_Settings.xml
Update to the mod settings.  Lots of old descriptions and tooltips that were literally just the name of the setting repeated.  This should be somewhat clearer for users to understand what each setting does.  Also clears up a few typos.

Just waiting on confirmation that the "DSU Item Destruction" thing is actually obsolete.  If it is, needs a matching patch somewhere else to remove the setting entirely.