Update SAL help text descriptions.
Removes all references to "deprecated" and just lists the SALs as "alternative" mechanisms for factory automation.  Also added help text for the SAL AutoCooker, AutoMiner, GodlyCrafter, and Overclocked Auto Assembler II, though I don't know if they're called in the appropriate place in the code or how that's set up for them to appear.

Each help text now also contains a couple sentences about the skill levels of the various SALs as well, so that it's more obvious how the different skill levels work.